feat: move logic to corresponding file
