+++ nodebuilder/p2p/routing.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/celestiaorg/celestia-node/nodebuilder/util"
 	"github.com/ipfs/go-datastore"
 	dht "github.com/libp2p/go-libp2p-kad-dht"
 	"github.com/libp2p/go-libp2p/core/protocol"
@@ -29,7 +28,7 @@ func peerRouting(cfg Config, params routingParams) (routing.PeerRouting, error) 
 		dht.RoutingTableRefreshPeriod(cfg.RoutingTableRefreshPeriod),
 	}
 
-	if util.IsBootstrapper {
+	if IsBootstrapper {
 		// override options for bootstrapper
 		opts = append(opts,
 			dht.Mode(dht.ModeServer), // it must accept incoming connections
